Move example dataset fixtures to conftest.py (#678)
* Move fixtures to conftest.py

* Revert "Move fixtures to conftest.py"

This reverts commit 3d02b15d

* Move examples dataset fixtures to conftest.py
