+++ tests/conftest.py
@@ -1,0 +1,32 @@
+from pytest import fixture
+from pyvista import examples
+
+
+@fixture()
+def airplane():
+    return examples.load_airplane()
+
+
+@fixture()
+def rectilinear():
+    return examples.load_rectilinear()
+
+
+@fixture()
+def sphere():
+    return examples.load_sphere()
+
+
+@fixture()
+def uniform():
+    return examples.load_uniform()
+
+
+@fixture()
+def ant():
+    return examples.load_ant()
+
+
+@fixture()
+def globe():
+    return examples.load_globe()
